pyproject.toml
[project]
name = "vkforge"
version = "0.1.0"
description = "VkForge: A Vulkan API Implementation Generator for Renderer Development"
authors = [
    { name = "Alrick Grandison", email = "alrickgrandison@gmail.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pydantic>=2.0",
    "PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/Rickodesea/VkForge"

[project.scripts]
vkforge = "vkforge.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"vkforge" = ["templates/*"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
